requirements.txt
astroid==2.2.5
attrs==19.1.0
autopep8==1.4.4
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.4.1
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.4
django-allauth==0.40.0
django-appconf==1.0.3
django-bootstrap-pagination==1.7.1
django-bootstrap4==1.0.1
django-extensions==2.2.3
django-heroku==0.3.1
django-imagekit==4.0.2
entrypoints==0.3
Faker==1.0.7
Flask==1.1.1
gunicorn==19.9.0
idna==2.8
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.14.1
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.4
oauthlib==3.1.0
pandas==0.25.0
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
pilkit==2.0
Pillow==6.2.1
pprint==0.1
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psycopg2==2.8.4
pycodestyle==2.5.0
Pygments==2.4.2
pylint==2.3.1
pyrsistent==0.15.3
python-dateutil==2.8.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.1
pywinpty==0.5.5
pyzmq==18.0.2
qtconsole==4.5.1
requests==2.22.0
requests-oauthlib==1.2.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.2
tornado==6.0.3
traitlets==4.3.2
typed-ast==1.4.0
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
whitenoise==4.1.4
widgetsnbextension==3.5.0
wrapt==1.11.2
